pyproject.toml
[project]
name = "datagouv-prospector-web"
version = "0.1.0"
description = "IA Prospector Web - Insee & Gemini"
readme = "INSTRUCTIONS.md"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.55.0",
    "pandas>=2.3.3",
    "plotly>=6.6.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "google-genai",
]

[tool.uv]
managed = true
